Fixed progress bar when there are too many steps. - Maybe will slow down a bit, but not that noticable.

+++ firmware/common/ui_widget.cpp
@@ -578,7 +578,7 @@ void ProgressBar::paint(Painter& painter) {
 	const auto sr = screen_rect();
 	const auto s = style();
 
-	v_scaled = (sr.size().width() * _value) / _max;
+	v_scaled = sr.size().width() * ((float)_value / (float)_max) ;
 
 	painter.fill_rectangle({sr.location(), {v_scaled, sr.size().height()}}, style().foreground);
 	painter.fill_rectangle({{sr.location().x() + v_scaled, sr.location().y()}, {sr.size().width() - v_scaled, sr.size().height()}}, s.background);
